pyproject.toml
[project]
name = "perplexity-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "flask>=3.1.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
